requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0.76
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
scikit-learn>=1.3.0
scipy>=1.10.0
transformers>=4.31.0
timm>=0.9.0
pytest>=7.4.0
Pillow>=9.5.0
